support range partition in shuffle splitter
