+++ utils/local-engine/Shuffle/ShuffleSplitter.h
@@ -6,7 +6,7 @@
 #include <IO/WriteBufferFromFile.h>
 #include <Common/PODArray.h>
 #include <Common/PODArray_fwd.h>
-
+#include <Shuffle/SelectorBuilder.h>
 
 //using namespace DB;
 
@@ -19,7 +19,8 @@ struct SplitOptions
     std::string local_tmp_dir;
     int map_id;
     size_t partition_nums;
-    std::vector<std::string> exprs;
+    std::string exprs;
+    // std::vector<std::string> exprs;
     std::string compress_method = "zstd";
     int compress_level;
 };
@@ -93,13 +94,13 @@ class RoundRobinSplitter : public ShuffleSplitter
 public:
     static std::unique_ptr<ShuffleSplitter> create(SplitOptions && options);
 
-    explicit RoundRobinSplitter(SplitOptions options_) : ShuffleSplitter(std::move(options_)) { }
+    explicit RoundRobinSplitter(SplitOptions options_);
 
     ~RoundRobinSplitter() override = default;
     void computeAndCountPartitionId(DB::Block & block) override;
 
 private:
-    int32_t pid_selection = 0;
+    std::unique_ptr<RoundRobinSelectorBuilder> selector_builder;
 };
 
 class HashSplitter : public ShuffleSplitter
@@ -107,15 +108,24 @@ class HashSplitter : public ShuffleSplitter
 public:
     static std::unique_ptr<ShuffleSplitter> create(SplitOptions && options);
 
-    explicit HashSplitter(SplitOptions options_) : ShuffleSplitter(std::move(options_)) { }
+    explicit HashSplitter(SplitOptions options_);
 
     ~HashSplitter() override = default;
     void computeAndCountPartitionId(DB::Block & block) override;
 
 private:
-    DB::FunctionBasePtr hash_function;
+    std::unique_ptr<HashSelectorBuilder> selector_builder;
 };
 
+class RangeSplitter : public ShuffleSplitter
+{
+public:
+    static std::unique_ptr<ShuffleSplitter> create(SplitOptions && options);
+    explicit RangeSplitter(SplitOptions options_);
+    void computeAndCountPartitionId(DB::Block & block) override;
+private:
+    std::unique_ptr<RangeSelectorBuilder> selector_builder;
+};
 struct SplitterHolder
 {
     ShuffleSplitter::Ptr splitter;
